fix: scaling issue in the regressor

+++ model/regressor_2Stage.py
@@ -135,8 +135,10 @@ class Regressor2Stage(nn.Module):
         inds = inds2.reshape((batches, 1, self.height, self.width))
         #print(id(inds))
         inds -= offset * self.stage_1_classes * self.stage_2_classes
+        #debug = torch.max(inds)
+        #print(debug)
         # the output lies between 0 and 1 to indicate the x position in the dot-pattern projector
-        x = (inds.type(torch.float) + x_2[:, 0, :, :]) * (1.0 / self.stage_1_classes * self.stage_2_classes)
+        x = (inds.type(torch.float) + x_2[:, 0, :, :]) * (1.0 / float(self.stage_1_classes * self.stage_2_classes))
 
         # one last relu for the masking
         mask = F.leaky_relu(x_2[:, 1, :, :])
